requirements.txt
mysql-connector-python
pandas
numpy
scikit-learn
matplotlib
seaborn
python-dotenv
